# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: rechvix API
  version: "0.1.0"
  description: >
    Multi-company, multi-branch, multi-warehouse billing, inventory,
    accounting, and India-GST tax-management API. This spec is generated
    by reading `internal/modules/*/httpapi/handlers.go` directly — every
    path, method, and request/response shape below is verified against
    the actual Go handler code, not inferred from `docs/architecture.md`'s
    design intent. See `docs/api.md` for narrative conventions (auth,
    error envelope, money shape, pagination status) that apply across
    every endpoint below rather than being repeated per-operation.

    Status: complete — all 124 real routes across all 15 HTTP modules are
    present and typed, mechanically verified against a grep of every
    r.Get/Post/Put/Delete/Patch(...) registration in the codebase (exact
    match, see the git history for this file). Kept current as endpoints
    ship (docs/TODO.md's cross-cutting checklist item) — re-run that same
    grep diff after adding or changing a route, don't just trust this
    note.
servers:
  - url: /api/v1
    description: Same origin as the web app; also reachable directly on the configured HTTP_PORT.
security:
  - sessionCookie: []
  - apiKey: []
paths:
...
  /reports/sales/summary:
    get:
      tags: [Reporting]
      summary: "Sales summary. Headers: Key, Documents, Taxable, Tax, Grand Total."
      parameters:
        - $ref: "#/components/parameters/ReportFilterFrom"
        - $ref: "#/components/parameters/ReportFilterTo"
        - $ref: "#/components/parameters/ReportFormat"
        - name: group_by
          in: query
          schema: { type: string, enum: [day, month, customer, product, category, salesperson, branch, warehouse], default: day }
      responses:
        "200": { $ref: "#/components/responses/ReportTableOK" }
        "400":
          description: INVALID_GROUP_BY.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /reports/sales/invoices:
    get:
      tags: [Reporting]
      summary: "Sales invoice detail. Headers: Document Number, Type, Customer, Issue Date, Status, Grand Total."
      parameters: [{ $ref: "#/components/parameters/ReportFilterFrom" }, { $ref: "#/components/parameters/ReportFilterTo" }, { $ref: "#/components/parameters/ReportFormat" }]
      responses: { "200": { $ref: "#/components/responses/ReportTableOK" } }
  /reports/sales/gross-profit:
    get:
      tags: [Reporting]
      summary: >
        Gross profit by product. Headers: Product, SKU, Qty Sold, Revenue,
        Approx COGS, Approx Profit. COGS is approximate — current
        average_cost, not the historical cost at the moment of sale (a
        known, documented limitation, docs/TODO.md Stage 7).
      parameters: [{ $ref: "#/components/parameters/ReportFilterFrom" }, { $ref: "#/components/parameters/ReportFilterTo" }, { $ref: "#/components/parameters/ReportFormat" }]
      responses: { "200": { $ref: "#/components/responses/ReportTableOK" } }
  /reports/purchases/summary:
    get:
      tags: [Reporting]
      summary: "Purchase summary. Headers: Key, Documents, Total. No taxable/tax breakdown yet — purchases isn't wired through the tax engine (known gap)."
      parameters:
        - $ref: "#/components/parameters/ReportFilterFrom"
        - $ref: "#/components/parameters/ReportFilterTo"
        - $ref: "#/components/parameters/ReportFormat"
        - { name: group_by, in: query, schema: { type: string, enum: [day, month, supplier, product, category, branch, warehouse], default: day } }
      responses: { "200": { $ref: "#/components/responses/ReportTableOK" } }
  /reports/purchases/documents:
    get:
      tags: [Reporting]
      summary: "Purchase document detail. Headers: Document Number, Type, Supplier, Date, Status, Total."
      parameters: [{ $ref: "#/components/parameters/ReportFilterFrom" }, { $ref: "#/components/parameters/ReportFilterTo" }, { $ref: "#/components/parameters/ReportFormat" }]
      responses: { "200": { $ref: "#/components/responses/ReportTableOK" } }
  /reports/inventory/valuation:
    get:
      tags: [Reporting]
      summary: "Stock valuation. Headers: Warehouse, Product, SKU, Qty On Hand, Avg Cost, Total Value."
      parameters: [{ name: warehouse_id, in: query, schema: { type: string, format: uuid } }, { $ref: "#/components/parameters/ReportFormat" }]
      responses: { "200": { $ref: "#/components/responses/ReportTableOK" } }
  /reports/inventory/low-stock:
    get:
      tags: [Reporting]
      summary: "Low stock. Headers: Warehouse, Product, SKU, Qty On Hand, Reorder Level."
      parameters: [{ name: warehouse_id, in: query, schema: { type: string, format: uuid } }, { $ref: "#/components/parameters/ReportFormat" }]
      responses: { "200": { $ref: "#/components/responses/ReportTableOK" } }
  /reports/inventory/movements:
    get:
      tags: [Reporting]
      summary: "Stock movements. Headers: Date, Warehouse, Product, SKU, Type, Quantity, Reference."
      parameters: [{ name: warehouse_id, in: query, schema: { type: string, format: uuid } }, { name: product_variant_id, in: query, schema: { type: string, format: uuid } }, { $ref: "#/components/parameters/ReportFormat" }]
      responses: { "200": { $ref: "#/components/responses/ReportTableOK" } }
  /reports/accounting/trial-balance:
    get:
      tags: [Reporting]
      summary: "Trial balance as of a date. Headers: Code, Account, Type, Debit, Credit."
      parameters: [{ name: as_of, in: query, schema: { type: string, format: date } }, { $ref: "#/components/parameters/ReportFormat" }]
      responses: { "200": { $ref: "#/components/responses/ReportTableOK" } }
  /reports/accounting/receivables:
    get:
      tags: [Reporting]
      summary: "Org-wide receivables ageing. Headers: Party ID, Current, 1-30, 31-60, 61-90, 90+, Total."
      parameters: [{ name: as_of, in: query, schema: { type: string, format: date } }, { $ref: "#/components/parameters/ReportFormat" }]
      responses: { "200": { $ref: "#/components/responses/ReportTableOK" } }
  /reports/accounting/payables:
    get:
      tags: [Reporting]
      summary: "Org-wide payables ageing. Same shape as receivables."
      parameters: [{ name: as_of, in: query, schema: { type: string, format: date } }, { $ref: "#/components/parameters/ReportFormat" }]
      responses: { "200": { $ref: "#/components/responses/ReportTableOK" } }
  /reports/accounting/accounts/{id}/ledger:
    get:
      tags: [Reporting]
      summary: "Account ledger (cash/bank book). Headers: Date, Source, Description, Debit, Credit, Balance."
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - $ref: "#/components/parameters/ReportFilterFrom"
        - $ref: "#/components/parameters/ReportFilterTo"
        - $ref: "#/components/parameters/ReportFormat"
      responses: { "200": { $ref: "#/components/responses/ReportTableOK" } }
  /reports/tax/hsn-summary:
    get:
      tags: [Reporting]
      summary: "HSN summary. Headers: HSN/SAC, Taxable, CGST, SGST, IGST, UTGST, CESS, Total Tax."
      parameters: [{ $ref: "#/components/parameters/ReportFilterFrom" }, { $ref: "#/components/parameters/ReportFilterTo" }, { $ref: "#/components/parameters/ReportFormat" }]
      responses: { "200": { $ref: "#/components/responses/ReportTableOK" } }
  /reports/tax/rate-summary:
    get:
      tags: [Reporting]
      summary: "Tax rate summary. Headers: Rate %, Taxable, Tax, Documents."
      parameters: [{ $ref: "#/components/parameters/ReportFilterFrom" }, { $ref: "#/components/parameters/ReportFilterTo" }, { $ref: "#/components/parameters/ReportFormat" }]
      responses: { "200": { $ref: "#/components/responses/ReportTableOK" } }
  /reports/tax/gstr1:
    get:
      tags: [Reporting]
      summary: >
        GSTR-1-oriented preparation — explicitly NOT a filing submission
        (title says so in the response too). Headers: Document Number,
        Date, Supply Type, Customer GSTIN, Place of Supply, Taxable,
        CGST, SGST, IGST, CESS, Grand Total. GSTR-3B is deliberately not
        built (docs/TODO.md Stage 7 — its inward/ITC side needs
        purchases' tax-engine wiring, which doesn't exist).
      parameters: [{ $ref: "#/components/parameters/ReportFilterFrom" }, { $ref: "#/components/parameters/ReportFilterTo" }, { $ref: "#/components/parameters/ReportFormat" }]
      responses: { "200": { $ref: "#/components/responses/ReportTableOK" } }

  # ============================================================
  # E-WAY BILL — free-first portal workflow (docs/architecture.md §9b,
  # Stage 8c). Mounted partly under /sales/documents/{id}/ewaybill/*
  # (document-scoped) and partly under /ewaybill (not document-scoped).
  # Verified against internal/modules/ewaybill/{domain,app,eligibility}.
  # ============================================================
...
  /webhooks/endpoints:
    post:
      tags: [Webhooks]
      summary: Register a webhook endpoint. The signing secret is shown exactly once (same principle as an API key's raw value).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, subscribed_events]
              properties:
                url: { type: string, format: uri }
                subscribed_events:
                  type: array
                  items: { type: string }
                  description: "Currently-wired producer events: invoice.finalized, einvoice.generated, einvoice.failed (docs/api.md — the brief's fuller event catalog isn't all wired yet)."
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  url: { type: string }
                  subscribed_events: { type: array, items: { type: string } }
                  signing_secret: { type: string, description: HMAC-SHA256 signing secret — store it now, it is never shown again. }
        "400":
          description: INVALID_REQUEST.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    get:
      tags: [Webhooks]
      summary: List webhook endpoints (signing secret never included).
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: object
                properties:
                  endpoints:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        url: { type: string }
                        subscribed_events: { type: array, items: { type: string } }
                        is_active: { type: boolean }
  /webhooks/endpoints/{id}:
    delete:
      tags: [Webhooks]
      summary: Deactivate a webhook endpoint.
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: object, properties: { status: { type: string, example: deactivated } } } } } }
components:
  parameters:
    ReportFilterFrom:
      name: from
      in: query
      schema: { type: string, format: date }
      description: YYYY-MM-DD.
    ReportFilterTo:
      name: to
      in: query
      schema: { type: string, format: date }
    ReportFormat:
      name: format
      in: query
      schema: { type: string, enum: [json, csv, xlsx, pdf], default: json }
      description: "csv/xlsx/pdf responses are attachments (Content-Disposition set); export is synchronous only — no background-job/expiring-link path yet (docs/api.md)."
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: bp_session
      description: Set by POST /auth/login. Default cookie name; configurable via SESSION_COOKIE_NAME.
    apiKey:
      type: http
      scheme: bearer
      description: >
        Authorization: Bearer <key>. Effective permissions are the
        intersection of the key's declared scopes and the owning user's
        real RBAC grants — see docs/api.md.
  responses:
    ReportTableOK:
      description: OK. For format=csv/xlsx/pdf, the response is the corresponding binary/text content type as an attachment instead.
      content:
        application/json: { schema: { $ref: "#/components/schemas/ReportTable" } }
        text/csv: { schema: { type: string, format: binary } }
        application/vnd.openxmlformats-officedocument.spreadsheetml.sheet: { schema: { type: string, format: binary } }
        application/pdf: { schema: { type: string, format: binary } }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string, example: SESSION_INVALID }
            message: { type: string }
            details: { type: object, additionalProperties: true }
            request_id: { type: string }
...
    ReportTable:
      type: object
      description: >
        internal/platform/export.Table's JSON shape — shared by every
        /reports/* listing endpoint. headers and each rows entry are
        ordered, positionally paired (rows[i][j] is the value for
        headers[j]) — see the WriteJSON doc comment for why this isn't a
        map-per-row shape.
      properties:
        title: { type: string }
        headers: { type: array, items: { type: string } }
        rows: { type: array, items: { type: array, items: { type: string } } }
...